reddit-crawler/requirements.txt
praw>=7.8.1
python-dotenv>=1.0.0
pandas==2.2.3